applied policies to most of the functions

+++ app/Http/Controllers/FileSystemEntryController.php
@@ -45,6 +45,7 @@ class FileSystemEntryController extends Controller
      */
     public function store(Request $request, $parent)
     {
+        $this->authorize('upload',$parent);
         $fse = FileSystemEntry::query()->create([
             'group_approval_id' => $request->get('group_approval_id'),
             'category_id' => $request->get('category'),
@@ -97,6 +98,7 @@ class FileSystemEntryController extends Controller
      */
     public function show(FileSystemEntry $fileSystemEntry)
     {
+        $this->authorize('view',$fileSystemEntry);
         return response($fileSystemEntry->loadMissing(['category', 'group_approval']));
     }
 
@@ -109,6 +111,7 @@ class FileSystemEntryController extends Controller
      */
     public function update(Request $request, FileSystemEntry $fileSystemEntry)
     {
+        $this->authorize('upload',$fileSystemEntry);
         $fileSystemEntry->name = $request->get('name');
         $fileSystemEntry->category_id = $request->get('category');
         $fileSystemEntry->group_approval_id = $request->get('group_approval_id');
@@ -125,6 +128,7 @@ class FileSystemEntryController extends Controller
      */
     public function destroy(FileSystemEntry $fileSystemEntry)
     {
+        $this->authorize('delete',$fileSystemEntry);
         if ($fileSystemEntry->is_directory) {
             $fileSystemEntry->descendantsAndSelf()->delete();
         } else {
@@ -168,12 +172,16 @@ class FileSystemEntryController extends Controller
 
     public function add_version(Request $request, FileSystemEntry $fileSystemEntry)
     {
+        $this->authorize('upload',$fileSystemEntry);
+
         $fileSystemEntry->addMediaFromRequest('attachment')->toMediaCollection();
         return response($fileSystemEntry->loadMissing('media'));
     }
 
     public function delete_version(Request $request, FileSystemEntry $fileSystemEntry, $version)
     {
+        $this->authorize('delete',$fileSystemEntry);
+
         $fileSystemEntry->media()->find($version)->delete();
         return response($fileSystemEntry->loadMissing('media'));
     }
